Add initial content to the HTML file
I've added the initial content to the web page,
especially to the upper section.

The content is mostly titles and text, with the
addition of an image, links and buttons.

This content will be occupying the space of the
sections for the moment, it will change according
to future needs and changes.