Add some format and style to the web page
I have added some format and styles to some
sections of the web page. Among those are the call
to action section, in which I added the division
with background color and a button, and the new
borders around the images.

I also added some more format and style, in
previous edits.